refactor: rename UIMembershipEmailInput to UiMembershipEmailInput
